refactor: simplify types structure and char class encoding
