--- src/encoder/types.ts
@@ -1,9 +1,0 @@
-/**
- * Encoded regex pattern with information about its type (atom, sequence)
- */
-export interface EncodedRegex {
-  precedence: EncodePrecedence;
-  pattern: string;
-}
-
-export type EncodePrecedence = 'atom' | 'sequence' | 'disjunction';
